+++ HISTORY.rst
@@ -143,6 +143,7 @@ Full Change List
 - Fixing hemizygous count display in fold-outs (#646).
 - Changing ClinVar link-out to VCV entry instead of coordinates (#693).
 - Adding unit test for clinvar Vue app (#692).
+- Moving clinvar Vue app (#711).
 
 ------
 v1.2.0